requirements.txt
flask==3.1.2
flask-sqlalchemy==3.1.1
click==8.3.1
flask-security-too==5.7.1
argon2-cffi==25.1.0
